+++ src/models/gmm_group.h
@@ -58,7 +58,7 @@ public:
      * @brief Get Offset added to covariance matrices for convergence
      * @return Offset added to covariance matrices for convergence
      */
-    float get_covarianceOffset() const;
+    double get_covarianceOffset() const;
     
     /**
      * @brief Set the number of mixture components of the model
@@ -74,7 +74,7 @@ public:
      * @param covarianceOffset offset to add to the diagonal of covariance matrices
      * @throws invalid_argument if the covariance offset is <= 0
      */
-    void set_covarianceOffset(float covarianceOffset);
+    void set_covarianceOffset(double covarianceOffset);
     
     /*@}*/
